requirements.txt
pdfplumber==0.11.0
pytest==7.4.3
pandas==2.1.4
llama-index-core==0.10.0
llama-index-llms-anthropic==0.1.0
llama-index-agent-openai==0.2.0
